Refactors DefaultDetailView to its own struct
